Included sender information in action information
It can be helpful to know about who sent the message/action. An example
of why this might be helpful is when using a content script to perform
something tab specific. Without the tab information you won't be able
to get the Tab ID or other useful information about where the action
came from.